LayoutTests/platform/blackberry/fast/dom/Document/early-document-access-expected.txt

CONSOLE MESSAGE: line 6: TypeError: Result of expression 'window.opener.pass' [undefined] is not a function.
FAIL: Timed out waiting for notifyDone to be called
